+++ index.html
@@ -50,8 +50,6 @@
       </p>
     </footer>
 
-    <script src="./index.js"></script>
-    <!-- <script src="./contentClass.js"></script> -->
-    <!-- <script src="spa.js"></script> -->
+    <script src="./index.js" type="module"></script>
   </body>
 </html>
