+++ modules/BookList.js
@@ -1,0 +1,10 @@
+export let books = [
+  {
+    author: 'ipsum',
+    title: 'Testyy',
+  },
+  {
+    author: 'lorem',
+    title: 'Testeroo',
+  },
+];
