+++ modules/localStorage.js
@@ -1,0 +1,5 @@
+// import { books } from './BookList.js';
+
+export const setLocalStore = (books) => {
+  localStorage.setItem('bookStore', JSON.stringify(books));
+};
